Make PathFileObject into an interface and hide the implementation
