+++ java-compiler-testing/src/main/java/io/github/ascopes/jct/workspaces/Workspaces.java
@@ -18,7 +18,6 @@ package io.github.ascopes.jct.workspaces;
 import io.github.ascopes.jct.compilers.JctCompiler;
 import io.github.ascopes.jct.utils.UtilityClass;
 import io.github.ascopes.jct.workspaces.impl.WorkspaceImpl;
-import java.util.Collection;
 import org.apiguardian.api.API;
 import org.apiguardian.api.API.Status;
