+++ java-compiler-testing/src/test/java/io/github/ascopes/jct/tests/unit/assertions/TraceDiagnosticListAssertTest.java
@@ -1208,7 +1208,6 @@ class TraceDiagnosticListAssertTest {
     }
   }
 
-  @SuppressWarnings("removal")
   @DisplayName("TraceDiagnosticListAssert#filteringBy tests")
   @Nested
   class FilteringByTest {
