+++ java-compiler-testing/src/test/java/io/github/ascopes/jct/tests/unit/workspaces/impl/JarFactoryImplTest.java
@@ -15,10 +15,12 @@
  */
 package io.github.ascopes.jct.tests.unit.workspaces.impl;
 
+import static org.assertj.core.api.Assertions.assertThat;
+import static org.assertj.core.api.Assertions.assertThatThrownBy;
+import static org.assertj.core.api.SoftAssertions.assertSoftly;
+
 import io.github.ascopes.jct.tests.helpers.Fixtures;
 import io.github.ascopes.jct.workspaces.impl.JarFactoryImpl;
-import org.junit.jupiter.api.DisplayName;
-import org.junit.jupiter.api.Test;
 import java.io.IOException;
 import java.nio.file.FileSystem;
 import java.nio.file.Files;
@@ -26,10 +28,8 @@ import java.nio.file.NoSuchFileException;
 import java.nio.file.Path;
 import java.nio.file.spi.FileSystemProvider;
 import java.util.Map;
-
-import static org.assertj.core.api.Assertions.assertThat;
-import static org.assertj.core.api.Assertions.assertThatThrownBy;
-import static org.assertj.core.api.SoftAssertions.assertSoftly;
+import org.junit.jupiter.api.DisplayName;
+import org.junit.jupiter.api.Test;
 
 /**
  * {@link JarFactoryImpl} tests.
